V0.0.5

+++ app/Clients/routes.php
@@ -1,8 +1,8 @@
 <?php
 Route::get('/clients', "ClientsController@index")->name('clients');
-Route::get('/clients/create', 'ClientsController@getCreate');
-Route::post('/clients/create', 'ClientsController@postCreate');
-Route::get('/clients/edit/{client}', 'ClientsController@getEdit');
-Route::post('/clients/edit/{client}', 'ClientsController@postEdit');
-Route::delete('/clients/delete/{client}', "ClientsController@deleteClient");
+Route::get('/clients/create', 'ClientsController@getCreate')->middleware('permissions');
+Route::post('/clients/create', 'ClientsController@postCreate')->middleware('permissions');
+Route::get('/clients/edit/{client}', 'ClientsController@getEdit')->middleware('permissions');
+Route::post('/clients/edit/{client}', 'ClientsController@postEdit')->middleware('permissions');
+Route::delete('/clients/delete/{client}', "ClientsController@deleteClient")->middleware('permissions');
 Route::post('clients/getProjects/{client}', "ClientsController@getProjectsByUsersWorkspaces");
